requirements.txt
tensorflow==2.15.0
streamlit==1.29.0
streamlit-drawable-canvas==0.9.3
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
Pillow==10.1.0